Verificaciones y checkAuth Multimedia

+++ controllers/Multimedia.controller.js
@@ -11,11 +11,13 @@ export const getAllMultimediaContents = async (req, res) => {
 
 export const createMultimediaContent = async (req, res) => {
  try {
+   if(!req.body.name || req.body.ref){
+      return res.status(400).json({ message:  'Por favor, ingrese todos los campos requeridos'  });
+   }
     const multimediaContent = new MultimediaContent({
       name: req.body.name,
       ref: req.body.ref,
     });
-
     await multimediaContent.save();
 
     res.status(201).json(multimediaContent);
@@ -27,6 +29,9 @@ export const createMultimediaContent = async (req, res) => {
 export const getMultimediaContent = async (req, res) => {
  try {
     const multimediaContent = await MultimediaContent.findById(req.params.id);
+    if(!multimediaContent){
+      return res.status(404).json({ message: 'MultimediaContent not found' });
+    }
     res.status(200).json(multimediaContent);
  } catch (error) {
     res.status(400).json({ message: error.message });
@@ -35,6 +40,13 @@ export const getMultimediaContent = async (req, res) => {
 
 export const updateMultimediaContent = async (req, res) => {
  try {
+   if(!req.body.name || req.body.ref){
+      return res.status(400).json({ message:  'Por favor, ingrese todos los campos requeridos'  });
+   }
+   const multimediaExist = await MultimediaContent.findById(req.params.id);
+   if(!multimediaExist){
+      return res.status(404).json({ message: 'MultimediaContent not found' });
+   }
     const multimediaContent = await MultimediaContent.findByIdAndUpdate(
       req.params.id,
       {
@@ -52,6 +64,10 @@ export const updateMultimediaContent = async (req, res) => {
 
 export const deleteMultimediaContent = async (req, res) => {
  try {
+   const multimediaContent = await MultimediaContent.findById(req.params.id);
+   if(!multimediaContent){
+      return res.status(404).json({ message: ' MultimediaContent not found' });
+   }
     await MultimediaContent.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: 'MultimediaContent deleted successfully' });
  } catch (error) {
